requirements.txt
fastapi
uvicorn[standard]
bcrypt
sqlalchemy
python-dotenv
pydantic[email]
python-jose[cryptography]
python-multipart
